+++ tests/test_visitors/test_ast/test_classes/test_methods/test_async_yield_magic_methods.py
@@ -1,8 +1,6 @@
-"""Test magic methods that can only be async generators or standard method."""
 import pytest
 
 from wemake_python_styleguide.violations.oop import AsyncMagicMethodViolation
-from wemake_python_styleguide.violations.oop import YieldMagicMethodViolation
 from wemake_python_styleguide.visitors.ast.classes import WrongMethodVisitor
 
 sync_method = """
@@ -18,37 +16,39 @@ class Example(object):
 """
 
 
+@pytest.mark.parametrize('template', [
+    sync_method,
+    async_method,
+])
 @pytest.mark.parametrize('method', [
     '__aiter__',
 ])
 @pytest.mark.parametrize('statement', [
     'yield',
     'yield 1',
-    'yield from some()',
 ])
-def test_wrong_sync_yield_magic_used(
+def test_yield_is_always_allowed_in_aiter(
     assert_errors,
-    assert_error_text,
     parse_ast_tree,
     default_options,
+    template,
     method,
     statement,
 ):
-    """Testing that the method cannot be a sync generator."""
-    tree = parse_ast_tree(sync_method.format(method, statement))
+    """Testing that the `__aiter__` can always have `yield`."""
+    tree = parse_ast_tree(template.format(method, statement))
 
     visitor = WrongMethodVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [YieldMagicMethodViolation])
-    assert_error_text(visitor, method)
+    assert_errors(visitor, [])
 
 
 @pytest.mark.parametrize('method', [
     '__aiter__',
 ])
 @pytest.mark.parametrize('statement', [
-    'print()',
+    'return some_async_iterator()',
 ])
 def test_wrong_async_magic_used(
     assert_errors,
@@ -95,7 +95,7 @@ def test_correct_async_yield_magic_used(
     '__aiter__',
 ])
 @pytest.mark.parametrize('statement', [
-    'print()',
+    'return some_async_iterator()',
 ])
 def test_correct_sync_magic_used(
     assert_errors,
@@ -106,6 +106,41 @@ def test_correct_sync_magic_used(
 ):
     """Testing that the method can be a normal method."""
     tree = parse_ast_tree(sync_method.format(method, statement))
+
+    visitor = WrongMethodVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [])
+
+
+# Examples:
+
+correct_nested_example = """
+class Some:
+    {0}def __aiter__(self):
+        async def inner():
+            yield 1
+        return inner()
+"""
+
+
+@pytest.mark.parametrize('example', [
+    correct_nested_example,
+])
+@pytest.mark.parametrize('mode', [
+    # We don't use `mode()` fixture here, because we have a nested func.
+    '',  # sync
+    'async ',
+])
+def test_correct_examples(
+    assert_errors,
+    parse_ast_tree,
+    default_options,
+    example,
+    mode,
+):
+    """Testing specific real-life examples that should be working."""
+    tree = parse_ast_tree(example.format(mode))
 
     visitor = WrongMethodVisitor(default_options, tree=tree)
     visitor.run()